beancount.el: Fix beancount-collect to correctly ignore entry at point
The old code collects all matches and then its callers remove the
entry found at point, because that entry is the one we're in the
process of completing so it's presumably not (yet) valid. However,
when the entry at point is actually already valid, it is incorrectly
removed from the possible completions.

Change `beancount-collect` so it ignores matches around point. This
gives the same result except when the match around point is also found
elsewhere, in which case the new code keeps that completion whereas
the old one incorrectly discarded it.