beancount.el: Rework imenu support and add tests
In org-mode the imemu entries are organized in a tree, however this
only allows to jump to leaf nodes, with no possibility to select
intermediate nodes.  To avoid this problem, follow what outline-mode
does and organize the imenu entries in a flat structure.  To make it
easier to select the entries in the minibuffer, the outline indicators
are removed from the node labels.