Update COPYING to GPLv3
Beancount is released under the GPLv2. Thus the GPLv2 was chosen for
the beancount-mode repository was it was split off from beancount
repository.  However, since beancount.el gained a license header, it
has been advertised as being released under the GPLv3-or-later and all
code later contributed to beancopunt.el was implicitly licencensed
under this license.  The main authors of the code in beancount.el
agree to license the code under the GPLv3-or-later.  I am the author
of the other other files and I am relicensing their content under the
GPLv3-or-later.