Allow completion of tags and links not on their own line
Previously only tags and links that existed on their own line within a
transaction were able to be completed. Now a link or tag can be
completed anywhere, including on the transaction directive line, and
including when multiple tags or links already exist on a given line.

This also updates the thing-at-point defintion for beancount-link to
remove the arbitrary length limit while keeping (improving) the
worst-case time complexity O(n) where n is the length of the link at
the point.