Don't insert space after completions
Previously there were certain completions (non-timestamped directives,
options, and timestamped directives) which would insert a space
afterward--different behavior from the rest of the completions.

Additionally this trailing space was unconditionally inserted into the
journal, even if the buffer already included the full
completion-with-trailing-space. This behavior was due to the space not
being included in the matching regexp, and thus not being included in
the (match-beginning) (match-end) bounds. If we do want to keep the
completion-adds-trailing-space behavior, the matching regexps should
be updated to include the optional trailing space.

Finally, allow completions on bare "^option " lines which do not yet
have an opening quote.