Use beancount-level-N faces instead of org-level-N
There is no guarantee org will be loaded when beancount-mode is
activated, and if absent, the org-level-N faces won't be available and
you get a torrent of 'invalid face' errors.

We could use outline-N faces instead (after all, if outline-minor-mode
is active, they will be available), but the ideal is to define our own
beancount-level-N faces and inherit them from outline-N instead. This
makes it easier to customize outline headings only in beancount-mode
buffers.