Change keybindings to respect Emacs conventions
Adjust the keybindings in `beancount-mode-map' (which is bound to C-c
by default) to respect the Emacs keybinding conventions: namely, keep
C-c LETTER reserved for user-defined bindings, and don't have a
binding that ends in C-g, since users expect C-g to cancel a sequence.

We accomplish this mostly by just changing all instances of C-c LETTER
to C-c C-LETTER, with two exceptions: `beancount-transaction-clear' is
moved from C-c C-g to C-c C-c, and `beancount-insert-prices' is moved
from C-c p to C-c C-i (because C-c C-p is already recommended in
README.org as the binding for `outline-previous-visible-heading', when
one is using `outline-minor-mode', and we don't want to shadow that).

Fixes #1.