Change sign of prefix argument for `beancount-insert-date`
New behaviour:
`C-u 1 M-RET` inserts a date 1 day in the *past*.
`C-u 7 M-RET` inserts a date 1 week in the *past*.
`C-u -1 M-RET` inserts a date 1 week in *future*.

I don't think I've ever added a future transaction. Mostly, when I use a
prefix arg, it's to enter a date in the past.

This new default is more convenient.